refactor: integrate reasoning loop into edrr
