+++ docs/specifications/edrr_reasoning_loop_integration.md
@@ -1,8 +1,8 @@
 ---
 author: ChatGPT
 date: 2024-07-01
-last_reviewed: 2024-07-01
-status: draft
+last_reviewed: 2025-07-16
+status: implemented
 tags:
   - specification
   - methodology
