+++ src/devsynth/application/orchestration/edrr_coordinator.py
@@ -4,10 +4,8 @@ from typing import Any, Callable, Dict, List, Optional
 
 from devsynth.logging_setup import DevSynthLogger
 from devsynth.methodology.base import Phase
-from devsynth.methodology.dialectical_reasoning import reasoning_loop
-from devsynth.methodology.edrr_coordinator import (
-    EDRRCoordinator as MethodologyEDRRCoordinator,
-)
+from devsynth.methodology.edrr import EDRRCoordinator as MethodologyEDRRCoordinator
+from devsynth.methodology.edrr import reasoning_loop
 
 logger = DevSynthLogger(__name__)
