+++ src/devsynth/methodology/__init__.py
@@ -1,5 +1,5 @@
 from .adhoc import AdHocAdapter
-from .dialectical_reasoning import reasoning_loop
+from .edrr import reasoning_loop
 from .kanban import KanbanAdapter
 from .milestone import MilestoneAdapter
 from .sprint import SprintAdapter
