+++ src/devsynth/methodology/edrr/__init__.py
@@ -1,0 +1,4 @@
+from .coordinator import EDRRCoordinator
+from .reasoning_loop import reasoning_loop
+
+__all__ = ["EDRRCoordinator", "reasoning_loop"]
