+++ tests/unit/methodology/test_edrr_coordinator.py
@@ -1,6 +1,7 @@
 """Tests for the methodology EDRR coordinator."""
 
-from devsynth.methodology.edrr_coordinator import EDRRCoordinator
+from devsynth.exceptions import ConsensusError
+from devsynth.methodology.edrr import EDRRCoordinator
 
 
 def test_automate_retrospective_review_summarizes_results() -> None:
@@ -14,3 +15,12 @@ def test_automate_retrospective_review_summarizes_results() -> None:
     summary = coordinator.automate_retrospective_review(raw, 2)
     assert summary["positives"] == ["good"]
     assert summary["sprint"] == 2
+
+
+def test_record_consensus_failure_logs(mocker) -> None:
+    """It delegates consensus failures to the logger."""
+    coordinator = EDRRCoordinator()
+    error = ConsensusError("no consensus")
+    spy = mocker.patch("devsynth.methodology.edrr.coordinator.log_consensus_failure")
+    coordinator.record_consensus_failure(error)
+    spy.assert_called_once()
